prism: split cmd/stats.go into 7 per-concern files
Pure structural refactor with no behaviour changes. Distributes the 2141-line
cmd/stats.go into focused files:

- stats.go: Cobra command registration, runStats router, parseSinceFlag,
  legacySentinel constant, modelCosts map
- stats_metrics.go: sessionMetrics struct + methods, collectMetrics,
  groupEventsByHarnessSessionID, collectModelMetrics, modelMetrics,
  computeTurnCost
- stats_render.go: renderIncarnationDetail, renderSessionDetail,
  renderSessionCompactTable, runStatsDetail, resolveSessionArg
- stats_aggregate.go: runStatsIncarnations, runStatsSummary,
  runStatsHistorical, runStatsSession
- stats_events.go: runStatsDoomLoops, runStatsDenials, runStatsAsks
- stats_model.go: runStatsModel, renderModelBreakdown
- stats_format.go: formatting helpers (formatTokenCount, formatCost,
  formatDurationLong, agentShortName, truncateStr, percentileFloat64,
  splitModel, formatAgentSummary, formatLatency)

stats_compare.go (prism stats compare, from #1130) is unchanged — it was
already a separate file.

Closes #1151
